firefly-832: Moved files into final location
   - file in visualize/ui/v2/ moved to visualize/ui/
   - removed deprecated files
