+++ src/firefly/js/visualize/ui/MouseReadPopoutAll.jsx
